Commit AQS Dual-lite version
This bootloader is meant to be loaded from the SD card, and
includes hard-coded calibrations so that the bootloader can go
directly to DDR3 memory. The calibration is done for a single-rank
DIMM, which is pretty close to the final configuration.

The u-boot image must be located at sector 2, and it is loaded
to 0x20000000 in DDR space.

The u-boot then copies a linux zImage with a DTB file concatenated
to the end from sector 256 on disk to location 0x10008000 in RAM,
and then jumps to it using the bootz subsystem.

So, the overall card layout is meant to be like this:

dd if=u-boot.imx of=/dev/sdb seek=2
dd if=zImage-plus-dtb of=/dev/sdb seek=256

The DTS file must be set correctly to have the AQS pin mappings
for the dual-lite.